move Getpass into its own file; go mod tidy
