+++ go.mod
@@ -1,5 +1,3 @@
 module github.com/lanrat/czds
 
 go 1.13
-
-require github.com/jschauma/getpass v0.2.1
